major changes applied
