+++ Practical-6/README.md
@@ -1,1 +1,13 @@
+# Pagination in React 🔥
 
+[Live Demo](https://react-ts-pr3.netlify.app/)
+
+## Fetures & uses 💡
+
+-   Pagination Implement in React Redux & React Hooks
+
+#### Screenshot
+
+<img src="https://i.ibb.co/KbQh1mx/Practical-3.png" alt="Alternate Text" width=400px />
+
+<img src="https://i.ibb.co/KbQh1mx/Practical-3.png" alt="Alternate Text" width=400px />
