+++ Practical-6/package.json
@@ -23,6 +23,9 @@
         "@babel/preset-typescript": "^7.16.7",
         "@types/react": "^17.0.39",
         "@types/react-dom": "^17.0.11",
+        "@types/react-redux": "^7.1.23",
+        "@types/redux": "^3.6.0",
+        "@types/redux-thunk": "^2.1.0",
         "@types/uuid": "^8.3.4",
         "@typescript-eslint/eslint-plugin": "^5.11.0",
         "@typescript-eslint/parser": "^5.11.0",
@@ -65,7 +68,6 @@
         "@reduxjs/toolkit": "^1.8.0",
         "@tailwindcss/aspect-ratio": "^0.4.0",
         "@tailwindcss/custom-forms": "^0.2.1",
-        "@types/react-redux": "^7.1.22",
         "@types/react-router-dom": "^5.3.3",
         "axios": "^0.26.0",
         "concurrently": "^7.0.0",
@@ -77,6 +79,7 @@
         "react-icons": "^4.3.1",
         "react-redux": "^7.2.6",
         "react-router-dom": "^6.2.1",
+        "redux-devtools-extension": "^2.13.9",
         "uuid": "^8.3.2"
     }
 }
