+++ Practical-6/src/Components/Users_PR_5_6/PaginationComp.tsx
@@ -1,0 +1,53 @@
+import * as React from 'react';
+
+export interface IAppProps {
+    page: number;
+    setPage: (page: number) => void;
+}
+
+export default function PaginationComp({page, setPage}: IAppProps) {
+    return (
+        <div>
+            <ul className="flex items-center justify-center">
+                <li>
+                    <button
+                        className="block py-2 px-3 ml-0 leading-tight text-gray-500 bg-white rounded-l-lg border border-gray-300 hover:bg-gray-100 hover:text-gray-700 "
+                        onClick={() => setPage(page > 1 ? page - 1 : page)}>
+                        <svg
+                            className="w-5 h-5"
+                            fill="currentColor"
+                            viewBox="0 0 20 20"
+                            xmlns="http://www.w3.org/2000/svg">
+                            <path
+                                fillRule="evenodd"
+                                d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
+                                clipRule="evenodd"></path>
+                        </svg>
+                    </button>
+                </li>
+                <li>
+                    <button className="py-2 px-3 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 ">
+                        {' '}
+                        {page}
+                    </button>
+                </li>
+                <li>
+                    <button
+                        className="block py-2 px-3 leading-tight text-gray-500 bg-white rounded-r-lg border border-gray-300 hover:bg-gray-100 hover:text-gray-700"
+                        onClick={() => setPage(page < 2 ? page + 1 : page)}>
+                        <svg
+                            className="w-5 h-5"
+                            fill="currentColor"
+                            viewBox="0 0 20 20"
+                            xmlns="http://www.w3.org/2000/svg">
+                            <path
+                                fillRule="evenodd"
+                                d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
+                                clipRule="evenodd"></path>
+                        </svg>
+                    </button>
+                </li>
+            </ul>
+        </div>
+    );
+}
