+++ Practical-6/src/Components/Webpack_PR2/Spinner.tsx
