+++ Practical-6/src/Pages/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {FC} from 'react';
-import UsersList from '../Components/Users/UserList';
+import UsersList from '../Components/Users_PR_5_6/UserList';
 
 const Home: FC = () => {
     return <UsersList />;
