+++ Practical-6/src/Pages/IpPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {FC} from 'react';
-import IPInfoCard from '../Components/Pr2_cmp/IPCard';
+import IPInfoCard from '../Components/Webpack_PR2/IPCard';
 
 const IPInfo: FC = () => {
     return (
