+++ Practical-6/src/Pages/TodoPage.tsx
@@ -1,5 +1,5 @@
 import React, {FC} from 'react';
-import TodoList from '../Components/Todo/TodoList';
+import TodoList from '../Components/Todo_PR_4_3/TodoList';
 
 const TodoPage: FC = () => {
     return (
