+++ Practical-6/src/Redux/Normal Redux/Constant/userConstant.ts
@@ -1,0 +1,3 @@
+export const USER_LIST_REQUEST = 'USER_LIST_REQUEST';
+export const USER_LIST_SUCCESS = 'USER_LIST_SUCCESS';
+export const USER_LIST_FAILURE = 'USER_LIST_FAILURE';
